Adiconado policies para verificar se o usuário pode realizar alguma ação.

+++ ArquitetandoJavaSwing/src/views/categories/List.java
@@ -12,6 +12,7 @@ import java.util.ArrayList;
 import javax.swing.ImageIcon;
 import javax.swing.JOptionPane;
 import javax.swing.table.DefaultTableModel;
+import policies.CategoryPolicy;
 import sources.Category;
 import views.main.ApplicationView;
 
@@ -30,6 +31,7 @@ public class List extends javax.swing.JInternalFrame {
         titleLabel.setFont(new Font(Font.SANS_SERIF, Font.PLAIN, 18));
 
         initSearchForm();
+        disableActions();
 
         ArrayList<Category> categories = new CategoriesController().list(searchParams());
         setTableData(tableData(categories), tableHeader());
@@ -41,6 +43,11 @@ public class List extends javax.swing.JInternalFrame {
         cb_searchStatus.addItem("Ativos");
         cb_searchStatus.addItem("Inativos");
         cb_searchStatus.setSelectedIndex(1);
+    }
+    
+    private void disableActions() {
+        btn_new.setEnabled(CategoryPolicy.canNew());
+        btn_searchSubmit.setEnabled(CategoryPolicy.canSearch());
     }
 
     private void setTableData(Object[][] tableData, String[] tableHeader) {
@@ -119,13 +126,21 @@ public class List extends javax.swing.JInternalFrame {
                 int col = table.columnAtPoint(evt.getPoint());
 
                 if (row >= 0 && col == 2) {
-                    int id = (int) table.getModel().getValueAt(row, 0);
-
-                    ApplicationView.changeInternalFrame(new Form(id));
+                    if (CategoryPolicy.canEdit()) {
+                        int id = (int) table.getModel().getValueAt(row, 0);
+                        
+                        ApplicationView.changeInternalFrame(new Form(id));
+                    } else {
+                        JOptionPane.showMessageDialog(null, "Você não tem permição para alterar as categorias.");
+                    }
                 } else if (row >= 0 && col == 3) {
-                    int id = (int) table.getModel().getValueAt(row, 0);
-
-                    destroyCategory(id);
+                    if (CategoryPolicy.canEdit()) {
+                        int id = (int) table.getModel().getValueAt(row, 0);
+                        
+                        destroyCategory(id);
+                    } else {
+                        JOptionPane.showMessageDialog(null, "Você não tem permição para excluir as categorias.");
+                    }
                 }
             }
         });
