+++ ArquitetandoJavaSwing/src/views/projects/List.java
@@ -11,6 +11,7 @@ import java.util.ArrayList;
 import javax.swing.ImageIcon;
 import javax.swing.JOptionPane;
 import javax.swing.table.DefaultTableModel;
+import policies.ProjectPolicy;
 import sources.Project;
 import views.main.ApplicationView;
 
@@ -27,6 +28,8 @@ public class List extends javax.swing.JInternalFrame {
         initComponents();
 
         titleLabel.setFont(new Font(Font.SANS_SERIF, Font.PLAIN, 18));
+        
+        btn_new.setEnabled(ProjectPolicy.canNew());
 
         loadProjects();
     }
@@ -109,13 +112,21 @@ public class List extends javax.swing.JInternalFrame {
                 int col = table.columnAtPoint(evt.getPoint());
 
                 if (row >= 0 && col == 5) {
-                    int id = (int) table.getModel().getValueAt(row, 0);
-
-                    ApplicationView.changeInternalFrame(new Form(id));
+                    if (ProjectPolicy.canEdit()) {
+                        int id = (int) table.getModel().getValueAt(row, 0);
+                        
+                        ApplicationView.changeInternalFrame(new views.categories.Form(id));
+                    } else {
+                        JOptionPane.showMessageDialog(null, "Você não tem permição para alterar os projetos.");
+                    }
                 } else if (row >= 0 && col == 6) {
-                    int id = (int) table.getModel().getValueAt(row, 0);
-
-                    destroyProject(id);
+                    if (ProjectPolicy.canDestroy()) {
+                        int id = (int) table.getModel().getValueAt(row, 0);
+                        
+                        destroyProject(id);
+                    } else {
+                        JOptionPane.showMessageDialog(null, "Você não tem permição para excluir os projetos.");
+                    }
                 }
             }
         });
